+++ controllers/operator/certs/cert_configurations.go
@@ -1,13 +1,17 @@
 package certs
 
 import (
+	"fmt"
+
 	mdbv1 "github.com/10gen/ops-manager-kubernetes/api/v1/mdb"
 	"github.com/10gen/ops-manager-kubernetes/controllers/operator/scale"
 )
 
 type Options struct {
-	// Name is the name of the resource.
-	Name string
+	// CertSecretName is the name of the secret which contains the certs.
+	CertSecretName string
+	// ResourceName is the name of the resource.
+	ResourceName string
 	// Replicas is the number of replicas.
 	Replicas int
 	// Namespace is the namepsace the resource is in.
@@ -26,7 +30,8 @@ type Options struct {
 // StandaloneConfig returns a function which provides all of the configuration options required for the given Standalone.
 func StandaloneConfig(mdb mdbv1.MongoDB) Options {
 	return Options{
-		Name:                         mdb.Name,
+		ResourceName:                 mdb.Name,
+		CertSecretName:               getCertNameWithPrefixOrDefault(mdb, mdb.Name),
 		Namespace:                    mdb.Namespace,
 		ServiceName:                  mdb.ServiceName(),
 		Replicas:                     1,
@@ -38,7 +43,8 @@ func StandaloneConfig(mdb mdbv1.MongoDB) Options {
 // ReplicaSetConfig returns a struct which provides all of the configuration options required for the given Replica Set.
 func ReplicaSetConfig(mdb mdbv1.MongoDB) Options {
 	return Options{
-		Name:                         mdb.Name,
+		ResourceName:                 mdb.Name,
+		CertSecretName:               getCertNameWithPrefixOrDefault(mdb, mdb.Name),
 		Namespace:                    mdb.Namespace,
 		Replicas:                     scale.ReplicasThisReconciliation(&mdb),
 		ServiceName:                  mdb.ServiceName(),
@@ -51,7 +57,8 @@ func ReplicaSetConfig(mdb mdbv1.MongoDB) Options {
 // ShardConfig returns a struct which provides all of the configuration options required for the given shard.
 func ShardConfig(mdb mdbv1.MongoDB, shardNum int, scaler scale.ReplicaSetScaler) Options {
 	return Options{
-		Name:                         mdb.ShardRsName(shardNum),
+		ResourceName:                 mdb.ShardRsName(shardNum),
+		CertSecretName:               getCertNameWithPrefixOrDefault(mdb, mdb.ShardRsName(shardNum)),
 		Namespace:                    mdb.Namespace,
 		Replicas:                     scale.ReplicasThisReconciliation(scaler),
 		ServiceName:                  mdb.ShardServiceName(),
@@ -63,7 +70,8 @@ func ShardConfig(mdb mdbv1.MongoDB, shardNum int, scaler scale.ReplicaSetScaler)
 // MongosConfig returns a struct which provides all of the configuration options required for the given Mongos.
 func MongosConfig(mdb mdbv1.MongoDB, scaler scale.ReplicaSetScaler) Options {
 	return Options{
-		Name:                         mdb.MongosRsName(),
+		ResourceName:                 mdb.MongosRsName(),
+		CertSecretName:               getCertNameWithPrefixOrDefault(mdb, mdb.MongosRsName()),
 		Namespace:                    mdb.Namespace,
 		Replicas:                     scale.ReplicasThisReconciliation(scaler),
 		ServiceName:                  mdb.ServiceName(),
@@ -75,7 +83,8 @@ func MongosConfig(mdb mdbv1.MongoDB, scaler scale.ReplicaSetScaler) Options {
 // ConfigSrvConfig returns a struct which provides all of the configuration options required for the given ConfigServer.
 func ConfigSrvConfig(mdb mdbv1.MongoDB, scaler scale.ReplicaSetScaler) Options {
 	return Options{
-		Name:                         mdb.ConfigRsName(),
+		ResourceName:                 mdb.ConfigRsName(),
+		CertSecretName:               getCertNameWithPrefixOrDefault(mdb, mdb.ConfigRsName()),
 		Namespace:                    mdb.Namespace,
 		Replicas:                     scale.ReplicasThisReconciliation(scaler),
 		ServiceName:                  mdb.ConfigSrvServiceName(),
@@ -83,4 +92,21 @@ func ConfigSrvConfig(mdb mdbv1.MongoDB, scaler scale.ReplicaSetScaler) Options {
 		additionalCertificateDomains: mdb.Spec.Security.TLSConfig.AdditionalCertificateDomains,
 		horizons:                     mdb.Spec.Connectivity.ReplicaSetHorizons,
 	}
+}
+
+// getCertNameWithPrefixOrDefault returns the name of the cert that will store certificates for the given resource.
+// this takes into account the tlsConfig.prefix option.
+func getCertNameWithPrefixOrDefault(mdb mdbv1.MongoDB, defaultName string) string {
+	tlsConfig := mdb.GetSecurity().TLSConfig
+	if tlsConfig != nil {
+		// use old behaviour if name is specified
+		if tlsConfig.SecretRef.Name != "" {
+			return tlsConfig.SecretRef.Name
+		}
+
+		if tlsConfig.SecretRef.Prefix != "" {
+			return fmt.Sprintf("%s-%s-cert", tlsConfig.SecretRef.Prefix, defaultName)
+		}
+	}
+	return defaultName + "-cert"
 }