+++ docker/mongodb-enterprise-tests/tests/tls/tls_permissions_default.py
@@ -6,27 +6,32 @@ from tests.opsmanager.om_ops_manager_https import create_mongodb_tls_certs
 
 
 @fixture(scope="module")
-def certs_secret(namespace: str, issuer: str):
-    return create_mongodb_tls_certs(issuer, namespace, "test-tls-base-rs", "certs")
+def certs_secret_prefix(namespace: str, issuer: str):
+    create_mongodb_tls_certs(
+        issuer, namespace, "test-tls-base-rs", "certs-test-tls-base-rs-cert"
+    )
+    return "certs"
 
 
 @fixture(scope="module")
-def replica_set(issuer_ca_configmap: str, namespace: str, certs_secret) -> MongoDB:
+def replica_set(
+    issuer_ca_configmap: str, namespace: str, certs_secret_prefix
+) -> MongoDB:
     resource = MongoDB.from_yaml(
         find_fixture("test-tls-base-rs.yaml"), namespace=namespace
     )
-    resource.configure_custom_tls(issuer_ca_configmap, certs_secret)
+    resource.configure_custom_tls(issuer_ca_configmap, certs_secret_prefix)
     return resource.create()
 
 
 @mark.e2e_replica_set_tls_default
-def test_replica_set(replica_set: MongoDB, namespace: str):
+def test_replica_set(replica_set: MongoDB):
 
     replica_set.assert_reaches_phase(Phase.Running, timeout=400)
 
 
 @mark.e2e_replica_set_tls_default
-def test_file_has_correct_permissions(replica_set: MongoDB, namespace: str):
+def test_file_has_correct_permissions(namespace: str):
     # We test that the permissions are as expected by executing the stat
     # command on all the pem files in the secrets/certs directory
     cmd = [